# swagger/swagger.json
{"consumes":["application/json"],"produces":["application/json"],"schemes":["http"],"swagger":"2.0","host":"localhost:3000","basePath":"/","info":{"description":"Noopy API for all front ends by Noopy","title":"Noopy API","version":"1.0.0","termsOfService":"https://www.teamwork.com/termsofservice","licence":{"Name":"BSD","url":"http://opensource.org/licenses/BSD-2-Clause"},"contact":{"email":"cmeichel@free.fr"}},"paths":{"/beer/":{"get":{"tags":["Beer"],"summary":"Get recipes","description":"Get all paginated recipes","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"limit","in":"query","description":"Number of recipes to return","type":"integer"},{"name":"page","in":"query","description":"Page to return","type":"integer"}]},"post":{"tags":["Beer"],"summary":"Add recipe","description":"Add a new recipe","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"recipe","in":"body","description":"recipe","required":true,"type":"string"}]}},"/beer/{recipeId}":{"delete":{"tags":["Beer"],"summary":"Delete recipe","description":"Delete a recipe","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"recipeId","in":"path","description":"Identifier of the recipe","required":true,"type":"string"}]},"put":{"tags":["Beer"],"summary":"Add recipe","description":"Update a recipe","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"recipeId","in":"path","description":"Identifier of the recipe","required":true,"type":"string"},{"name":"recipe","in":"body","description":"recipe","required":true,"type":"string"}]}},"/beer/{recipeId}/step":{"delete":{"tags":["Beer"],"summary":"Delete step","description":"Delete a step from a recipe","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"recipeId","in":"path","description":"Identifier of the recipe","required":true,"type":"string"}]},"post":{"tags":["Beer"],"summary":"Add step","description":"Add a new step to the recipe","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"recipeId","in":"path","description":"Identifier of the recipe","required":true,"type":"string"},{"name":"step","in":"body","description":"Step","required":true,"type":"string"}]}},"/me/":{"get":{"tags":["Me"],"summary":"Read","description":"Get informatoin about current user","externalDocs":{},"consumes":["application/json"]}},"/quizz/":{"get":{"tags":["Quizz"],"summary":"Read","description":"Get Random questions","externalDocs":{},"consumes":["application/json"],"parameters":[{"name":"limit","in":"query","description":"Max questions","type":"integer"}]}}},"definitions":{},"tags":[{"Name":"Beer","description":"Beer resource resource","externalDocs":{}},{"Name":"Me","description":"Information about current User","externalDocs":{}},{"Name":"Quizz","description":"Roller Derby Quizz resource","externalDocs":{}}],"externalDocs":{},"securityDefinitions":{}}